+++ retz-client/src/main/java/io/github/retz/cli/CommandGetJob.java
@@ -51,7 +51,11 @@ public class CommandGetJob implements SubCommand {
     public int handle(FileConfiguration fileConfig) {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
-        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
+        try (Client webClient = Client.newBuilder(fileConfig.getUri())
+                .enableAuthentication(fileConfig.authenticationEnabled())
+                .setAuthenticator(fileConfig.getAuthenticator())
+                .checkCert(fileConfig.checkCert())
+                .build()) {
 
             LOG.info("Fetching job detail id={}", id);