+++ retz-client/src/main/java/io/github/retz/cli/CommandKill.java
@@ -47,7 +47,11 @@ public class CommandKill implements SubCommand {
     public int handle(FileConfiguration fileConfig) {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
-        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
+        try (Client webClient = Client.newBuilder(fileConfig.getUri())
+                .enableAuthentication(fileConfig.authenticationEnabled())
+                .setAuthenticator(fileConfig.getAuthenticator())
+                .checkCert(fileConfig.checkCert())
+                .build()) {
 
             LOG.info("Killing job detail id={}", id);
             Response res = webClient.kill(id);