+++ retz-client/src/main/java/io/github/retz/cli/CommandSchedule.java
@@ -74,7 +74,12 @@ public class CommandSchedule implements SubCommand {
                 envProps, parseRange(cpu, "1"), parseRange(mem, "32"), Integer.parseInt(gpu));
         job.setTrustPVFiles(trustPVFiles);
 
-        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
+        try (Client webClient = Client.newBuilder(fileConfig.getUri())
+                .enableAuthentication(fileConfig.authenticationEnabled())
+                .setAuthenticator(fileConfig.getAuthenticator())
+                .checkCert(fileConfig.checkCert())
+                .build()) {
+
             LOG.info("Sending job {} to App {}", job.cmd(), job.appid());
             Response res = webClient.schedule(job);
             if (res instanceof ScheduleResponse) {