+++ retz-client/src/main/java/io/github/retz/cli/CommandUnloadApp.java
@@ -47,7 +47,11 @@ public class CommandUnloadApp implements SubCommand {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
 
-        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
+        try (Client webClient = Client.newBuilder(fileConfig.getUri())
+                .enableAuthentication(fileConfig.authenticationEnabled())
+                .setAuthenticator(fileConfig.getAuthenticator())
+                .checkCert(fileConfig.checkCert())
+                .build()) {
 
             Response res = webClient.unload(appName);
             LOG.info(res.status());