+++ retz-common/src/test/java/io/github/retz/cli/FileConfigurationTest.java
@@ -16,13 +16,19 @@
  */
 package io.github.retz.cli;
 
+import io.github.retz.auth.Authenticator;
 import org.junit.Test;
 
+import javax.crypto.Mac;
+import javax.crypto.spec.SecretKeySpec;
 import java.io.File;
 import java.net.URI;
+import java.util.Base64;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 /**
  * Created by kuenishi on 9/14/16.
@@ -37,5 +43,39 @@ public class FileConfigurationTest {
         assertEquals(config.getKeystoreFile(), "path/to/keystore.jsk");
         assertEquals(config.getKeystorePass(), "foobar");
         assertFalse(config.checkCert());
+        assertTrue(config.authenticationEnabled());
+    }
+
+    @Test
+    public void signature() throws  Exception {
+        FileConfiguration config = new FileConfiguration("src/test/resources/retz.properties");
+
+        assertTrue(config.authenticationEnabled());
+        Authenticator a = config.getAuthenticator();
+        String verb = "GET";
+        String md5 = "1234567890";
+        String date = TimestampHelper.now();
+        String resource = "/job/234";
+        String signature = sign(verb, md5, date, resource, "deadbeef");
+        System.err.println("Signature: "+signature);
+        assertTrue(a.authenticate(verb, md5, date, resource, "cafebabe", signature));
+
+        String wrongSignature = sign(verb, md5, date, resource, "my name is charlie");
+        System.err.println("Wrong signature: "+wrongSignature);
+        assertFalse(a.authenticate(verb, md5, date, resource, "cafebabe", wrongSignature));
+    }
+
+    private String sign(String verb, String md5, String date, String resource, String secret) throws Exception {
+    String string2sign = new StringBuilder()
+            .append(verb).append("\n")
+            .append(md5).append("\n")
+            .append(date).append("\n")
+            .append(resource).append("\n")
+            .toString();
+        Mac mac = Mac.getInstance("HmacSHA256");
+
+        mac.init(new SecretKeySpec(secret.getBytes(UTF_8), "HmacSHA256"));
+
+        return Base64.getEncoder().withoutPadding().encodeToString(mac.doFinal(string2sign.getBytes(UTF_8)));
     }
 }